requirements.txt
asgiref==3.7.2
attrs==23.1.0
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drfpasswordless==1.5.9
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.15.0
hyperframe==6.0.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pipdeptree==2.13.0
pluggy==1.3.0
priority==2.0.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.6.0
pytz==2023.3.post1
PyYAML==6.0.1
rpds-py==0.10.6
referencing==0.30.2
requests==2.31.0
simplejson==3.19.2
tzdata==2023.3
uritemplate==4.1.1
wsproto==1.2.0
